requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
propcache==0.2.1
aiohttp==3.11.11